backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
torch>=1.12.0
transformers>=4.20.0
sentence-transformers>=2.2.2
requests>=2.28.0
python-dotenv>=0.20.0
safetensors>=0.3.0
tqdm>=4.64.0
